move helper funcs to util/deployment.go from util.go
